# docs/openapi.yml
openapi: 3.1.0
info:
  version: 1.1.0
  title: xButler webserver
  description: Web server for ML services of xButler

paths:
  /retrieval/text:
    get:
      parameters:
        - explode: true
          in: query
          name: text
          required: true
          schema:
            title: Text
            type: string
          style: form
        - explode: true
          in: query
          name: topk
          required: false
          schema:
            default: 2
            title: Topk
            type: integer
            format: int32
            minimum: 0
          style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/NearestItem"
                type: array
          description: Successful Response
      summary: Get top k resources via query text
      tags:
        - Retrieval
  /retrieval/image:
    post:
      parameters:
        - explode: true
          in: query
          name: topk
          required: false
          schema:
            default: 2
            title: Topk
            type: integer
            format: int32
            minimum: 0
          style: form
      requestBody:
        content:
          image/png:
            schema:
              format: binary
              type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/NearestItem"
                type: array
          description: Successful Response
      summary: Get top k similar resources via queried image
      tags:
        - Retrieval
  /resource/{resourceId}:
    get:
      parameters:
        - explode: false
          in: path
          name: resourceId
          description: Resource ID
          required: true
          schema:
            type: string
            format: uuid

      responses:
        "200":
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            video/mp4:
              schema:
                type: string
                format: binary
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
          description: Validation Error
      summary: Get resource via ID
      tags:
        - Resource

  /resource:
    post:
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/Body_upload_item_api_resource_put"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
          description: Validation Error
      summary: Upload item
      tags:
        - Resource
components:
  schemas:
    Body_upload_item_api_resource_put:
      properties:
        file:
          format: binary
          title: File
          type: string
      required:
        - file
      title: Body_upload_item_api_resource_put
    HTTPError: {}
    NearestItem:
      example:
        item_id: 0
      properties:
        item_id:
          title: Item Id
          type: integer
      required:
        - item_id
      title: NearestItem